+++ test/number_to_words_nepali_test.dart
@@ -3,13 +3,13 @@ import 'package:number_to_words_nepali/number_to_words_nepali.dart';
 
 void main() {
   group('Nepali Language', () {
-    final Map<int, String> testNumbers = {
-      10000: 'दस हजार',
-      100001: 'एक लाख एक',
-      11111111: 'एक करोड एघार लाख एघार हजार एक सय एघार',
-      123456789: 'बाह्र करोड चौतीस लाख छपन्न हजार सात सय उनान्नब्बे',
+    final Map<dynamic, String> testNumbers = {
+      10000: 'दश हजार',
+      '100001': 'एक लाख एक',
+      111111.11: 'एक लाख एघार हजार एक सय एघार दशमलव एक एक',
+      '1234567.89': 'बाह्र लाख चौँतीस हजार पाँच सय सतसट्ठी दशमलव आठ नौ',
       999999999999999999:
-          'नौ शंख उनान्सय पद्म उनान्सय नील उनान्सय खरब उनान्सय अरब उनान्सय करोड उनान्सय लाख उनान्सय हजार नौ सय उनान्सय'
+          'नौ शंख उनान्सय पद्म उनान्सय नील उनान्सय खर्ब उनान्सय अर्ब उनान्सय करोड उनान्सय लाख उनान्सय हजार नौ सय उनान्सय'
     };
 
     for (var testNumber in testNumbers.entries) {
@@ -20,13 +20,34 @@ void main() {
     }
   });
 
+  group('Nepali Language Monetary', () {
+    final Map<dynamic, String> testNumbers = {
+      10000: 'दश हजार रुपैंया',
+      '100001': 'एक लाख एक रुपैंया',
+      111111.11: 'एक लाख एघार हजार एक सय एघार रुपैंया, एघार पैसा',
+      '1234567.89':
+          'बाह्र लाख चौँतीस हजार पाँच सय सतसट्ठी रुपैंया, उनान्नब्बे पैसा',
+      999999999999999999:
+          'नौ शंख उनान्सय पद्म उनान्सय नील उनान्सय खर्ब उनान्सय अर्ब उनान्सय करोड उनान्सय लाख उनान्सय हजार नौ सय उनान्सय रुपैंया'
+    };
+
+    for (var testNumber in testNumbers.entries) {
+      test('${testNumber.key}: ${testNumber.value}', () {
+        expect(
+            NumberToWordsNepali(isMonetary: true)
+                .convertNumberToWordsNepali(testNumber.key),
+            testNumber.value);
+      });
+    }
+  });
+
   group('English Language', () {
-    final Map<int, String> testNumbers = {
+    final Map<dynamic, String> testNumbers = {
       10000: 'ten thousand',
-      100001: 'one lakh one',
-      11111111: 'one crore eleven lakh eleven thousand one hundred eleven',
-      123456789:
-          'twelve crore thirty-four lakh fifty-six thousand seven hundred eighty-nine',
+      '100001': 'one lakh one',
+      111111.11: 'one lakh eleven thousand one hundred eleven point one one',
+      '1234567.89':
+          'twelve lakh thirty-four thousand five hundred sixty-seven point eight nine',
       999999999999999999:
           'nine shankh ninety-nine padma ninety-nine neel ninety-nine kharab ninety-nine arab ninety-nine crore ninety-nine lakh ninety-nine thousand nine hundred ninety-nine'
     };
@@ -35,6 +56,49 @@ void main() {
       test('${testNumber.key}: ${testNumber.value}', () {
         expect(
             NumberToWordsNepali(language: Language.english)
+                .convertNumberToWordsNepali(testNumber.key),
+            testNumber.value);
+      });
+    }
+  });
+
+  group('English Language Monetary', () {
+    final Map<dynamic, String> testNumbers = {
+      10000: 'ten thousand rupees',
+      '100001': 'one lakh one rupees',
+      111111.11:
+          'one lakh eleven thousand one hundred eleven rupees and eleven paisa',
+      '1234567.89':
+          'twelve lakh thirty-four thousand five hundred sixty-seven rupees and eighty-nine paisa',
+      999999999999999999:
+          'nine shankh ninety-nine padma ninety-nine neel ninety-nine kharab ninety-nine arab ninety-nine crore ninety-nine lakh ninety-nine thousand nine hundred ninety-nine rupees'
+    };
+
+    for (var testNumber in testNumbers.entries) {
+      test('${testNumber.key}: ${testNumber.value}', () {
+        expect(
+            NumberToWordsNepali(language: Language.english, isMonetary: true)
+                .convertNumberToWordsNepali(testNumber.key),
+            testNumber.value);
+      });
+    }
+  });
+
+  group('Error Handling', () {
+    final Map<dynamic, Matcher> testNumbers = {
+      '': throwsRangeError,
+      '.': throwsArgumentError,
+      '-.': throwsArgumentError,
+      'NaN': throwsArgumentError,
+      '+.12e-9': throwsArgumentError,
+      '99999999999999999999999999': throwsRangeError,
+      '9999999999999999999.999999999999999': throwsRangeError,
+    };
+
+    for (var testNumber in testNumbers.entries) {
+      test('${testNumber.key}', () {
+        expect(
+            () => NumberToWordsNepali()
                 .convertNumberToWordsNepali(testNumber.key),
             testNumber.value);
       });