+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.1.0
+* Added monetary conversion feature.
+* Implemented letter case conversion.
+* Now accepts int, double and string as input.
+* Updated lower Dart constraint to '>=2.19.0'.
+
 # 1.0.1
 * Updated documentation for various components.
 * Updated SDK constraint to >=2.19.5 <4.0.0.